handle interrupts while polling and sleeping
If a sleep or poll function returns EINTR than it
was interrupted by a signal (e.g. SIGPROF if the
application is beeing profiled). In that case one
should continue polling/sleeping.

Use clock_nanosleep instead of nanosleep to avoid
a time drift when the sleep is frequently interrupted,
which again is the case when using a profiler.