cmake: bump cmake_minimum_required()
We are using VERSION_GREATER_EQUAL [1] which was added in CMake 3.7
(released in November 2016). The distributions that we perform releases
for are all running at least this version, further substantiated by the
fact that we've used this feature since libfido2 1.9.0 (released in
October 2021).

Moreover, this change silences a warning on modern CMake versions that
compatilibity with versions less than CMake 3.5 is going to be removed
in a future CMake release [2].

[1] https://cmake.org/cmake/help/latest/command/if.html#version-greater-equal
[2] https://cmake.org/cmake/help/latest/command/cmake_minimum_required.html#policy-settings

This also resolves #718.